# stackhawk-openapi.yaml
openapi: 3.0.3
info:
  title: Options Copilot API
  version: "1.0.0"
  description: >
    Hand-derived from apps/api/src/app.ts and apps/api/src/practice.ts for HawkScan
    (no OpenAPI-generation tooling is wired into the Fastify app). Every route in
    apps/api/src/app.ts spends only what the caller's own connected wallet signs
    client-side (ADR-0011) -- /fill/prepare returns unsigned calldata, nothing here
    signs or submits a transaction, so nothing needs to be excluded from the scan.
servers:
  - url: http://127.0.0.1:3001
paths:
  /health:
    get:
      operationId: getHealth
      responses:
        "200":
          description: OK
  /book:
    get:
      operationId: getBook
      responses:
        "200":
          description: OK
  /session:
    get:
      operationId: getSession
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
  /session/budget:
    post:
      operationId: postSessionBudget
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [riskBudgetUsdc]
              properties:
                riskBudgetUsdc:
                  type: number
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
  /deck:
    get:
      operationId: getDeck
      parameters:
        - name: direction
          in: query
          required: true
          schema:
            type: string
            enum: [UP, DOWN]
        - name: horizonDays
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 3
        - name: sizeUsdc
          in: query
          required: true
          schema:
            type: number
            minimum: 0.01
            maximum: 1000
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
  /propose:
    post:
      operationId: postPropose
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [underlying, direction, sizeUsdc, horizonDays]
              properties:
                underlying:
                  type: string
                  enum: [ETH]
                direction:
                  type: string
                  enum: [UP, DOWN]
                sizeUsdc:
                  type: number
                  minimum: 0.01
                  maximum: 1000
                horizonDays:
                  type: integer
                  minimum: 1
                  maximum: 7
                cardRef:
                  type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "410":
          description: Gone
        "502":
          description: Bad Gateway
  /auth/challenge:
    post:
      operationId: postAuthChallenge
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [walletAddress]
              properties:
                walletAddress:
                  type: string
                  pattern: "^0x[a-fA-F0-9]{40}$"
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
  /auth/verify:
    post:
      operationId: postAuthVerify
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [signature]
              properties:
                signature:
                  type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "410":
          description: Gone
  /fill/prepare:
    post:
      operationId: postFillPrepare
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [proposalId, walletAddress]
              properties:
                proposalId:
                  type: string
                walletAddress:
                  type: string
                  pattern: "^0x[a-fA-F0-9]{40}$"
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "410":
          description: Gone
        "502":
          description: Bad Gateway
  /fill/settle:
    post:
      operationId: postFillSettle
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [proposalId]
              properties:
                proposalId:
                  type: string
                txHash:
                  type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "410":
          description: Gone
        "425":
          description: Too Early
        "502":
          description: Bad Gateway
  /positions:
    get:
      operationId: getPositions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
  /practice:
    post:
      operationId: postPractice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [proposalId]
              properties:
                proposalId:
                  type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "410":
          description: Gone
  /forecast/news:
    get:
      operationId: getForecastNews
      security:
        - bearerAuth: []
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: horizon
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "502":
          description: Bad Gateway
  /forecast/price:
    get:
      operationId: getForecastPrice
      security:
        - bearerAuth: []
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: horizon
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "502":
          description: Bad Gateway
  /forecast/risk-benefit:
    get:
      operationId: getForecastRiskBenefit
      security:
        - bearerAuth: []
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: horizon
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "502":
          description: Bad Gateway
  /forecast/indicators:
    get:
      operationId: getForecastIndicators
      security:
        - bearerAuth: []
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "502":
          description: Bad Gateway
        "503":
          description: Service Unavailable
  /forecast/ask:
    post:
      operationId: postForecastAsk
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question:
                  type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "502":
          description: Bad Gateway
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer